backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
motor==3.3.2
pymongo==4.6.0
python-dotenv==1.0.0
ultralytics==8.1.0
opencv-python==4.8.1.78
numpy==1.24.3
torch==2.1.1
torchvision==0.16.1
Pillow==10.1.0
scipy==1.11.4
slowapi==0.1.9
python-jose[cryptography]==3.3.0

